refactors list command into lib via trait generic specialization
Adds a CRUDable trait (name subject to change) that defines methods
intended to run Create, Read, Update, Delete type methods. It also
takes a generic of the Database type in order to be able to specialize
the methods for other types of databases in the future. For instance,
the "create" method will need to be specialized because the different
sqlx database implementations all have a different method for obtaining the
last insert id.